Fix casting and struct inits issues.
Former-commit-id: ba3ca1b3ff831c1b4ba568aa2274fb3ba2199c14

+++ src/Gbuffer.cpp
@@ -121,7 +121,7 @@ void Gbuffer::resize(int width, int height){
 }
 
 void Gbuffer::resize(glm::vec2 size){
-	resize(size[0],size[1]);
+	resize((int)size[0], (int)size[1]);
 }
 
 void Gbuffer::clean() const {
